fixes #15

+++ src/backend/routes/chat.go
@@ -11,6 +11,7 @@ import (
 	"os"
 	"os/exec"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/gorilla/mux"
@@ -23,7 +24,7 @@ import (
 // Initialize a Redis client as a global variable.
 var redisClient *redis.Client
 var ctx = context.Background()
-var chatFetchCmds = make(map[string]*exec.Cmd)
+var chatFetchCmds sync.Map // [string]*exec.Cmd
 
 var upgrader = websocket.Upgrader{
 	CheckOrigin: func(r *http.Request) bool {
@@ -133,7 +134,7 @@ func StartChatFetch(urls []string) {
 func monitorAndRestartChatFetch(url, pythonExecPath, fetchChatScript string) {
 	for {
 		cmd := startChatFetch(url, pythonExecPath, fetchChatScript)
-		chatFetchCmds[url] = cmd
+		chatFetchCmds.Store(url, cmd)
 
 		err := cmd.Wait() // Waits for the command to exit
 		if err != nil {
@@ -336,14 +337,20 @@ func ImageProxy(w http.ResponseWriter, r *http.Request) {
 
 // StopChatFetches stops all ongoing chat fetch commands
 func StopChatFetches(w http.ResponseWriter, r *http.Request) {
-	for _, cmd := range chatFetchCmds {
+	chatFetchCmds.Range(func(key, value any) bool {
+		cmd, ok := value.(*exec.Cmd)
+		if !ok {
+			log.Printf("Failed to coerce chatFetchCmd to type *exec.Cmd")
+			return true
+		}
 		if cmd != nil && cmd.Process != nil {
 			err := cmd.Process.Kill()
 			if err != nil {
 				log.Printf("Failed to stop chat fetch command: %v", err)
 			}
 		}
-	}
+		return true
+	})
 	fmt.Fprintln(w, "Chat fetch commands stopped. Restarting...")
 }
